+++ src/components/AbstractSelectableComponent.js
@@ -47,7 +47,7 @@ export default class AbstractSelectableComponent extends React.Component {
     }
 
     const { zoom } = this.getInitialViewState();
-    const editHandlePointRadius = 60 / (zoom + 16);
+    const editHandlePointRadius = 5 / (zoom + 16);
 
     return [new SelectionLayer({
       id: 'selection',
@@ -145,7 +145,7 @@ export default class AbstractSelectableComponent extends React.Component {
       };
     }
     return (
-      <React.Fragment>
+      <>
         <div className="d-flex">
           <ToolMenu {...toolProps} />
           {this.renderLayersMenu()}
@@ -158,7 +158,7 @@ export default class AbstractSelectableComponent extends React.Component {
         >
           {this.initializeViewInfo}
         </DeckGL>
-      </React.Fragment>
+      </>
     );
   }
 }