+++ src/components/sets/PopoverMenu.js
@@ -1,8 +1,6 @@
 import React, { useState } from 'react';
-import { Popover } from 'antd';
-import 'antd/es/popover/style/index.css';
-
 import { callbackOnKeyPress } from './utils';
+import Popover from './Popover';
 
 function PopoverMenuList(props) {
   const { menuConfig, onClick } = props;
@@ -46,7 +44,6 @@ export default function PopoverMenu(props) {
 
   return (
     <Popover
-      overlayClassName="vitessce-popover"
       content={<PopoverMenuList menuConfig={menuConfig} onClick={closePopover} />}
       placement={placement}
       trigger="click"