+++ src/components/sets/TreeNode.js
@@ -7,9 +7,9 @@ import PopoverMenu from './PopoverMenu';
 import PopoverColor from './PopoverColor';
 import { callbackOnKeyPress, range, levelNameFromIndex } from './utils';
 
-import EyeSVG from '../../assets/tools/eye.svg';
-import PenSVG from '../../assets/tools/pen.svg';
-import TrashSVG from '../../assets/tools/trash.svg';
+import { ReactComponent as EyeSVG } from '../../assets/tools/eye.svg';
+import { ReactComponent as PenSVG } from '../../assets/tools/pen.svg';
+import { ReactComponent as TrashSVG } from '../../assets/tools/trash.svg';
 
 function makeNodeViewMenuConfig(props) {
   const {
@@ -50,7 +50,7 @@ function NamedSetNodeStatic(props) {
   } = props;
   const [iconsVisible, setIconsVisible] = useState(false);
   return (
-    <React.Fragment>
+    <>
       <button
         type="button"
         onClick={() => { tree.viewSet(setKey); }}
@@ -90,7 +90,7 @@ function NamedSetNodeStatic(props) {
           <Icon component={TrashSVG} className={`${prefixClass}-node-menu-trigger`} title="Delete" />
         </PopoverMenu>
       </span>
-    </React.Fragment>
+    </>
   );
 }
 
@@ -103,7 +103,7 @@ function NamedSetNodeEditing(props) {
   } = props;
   const [currentTitle, setCurrentTitle] = useState(title);
   return (
-    <React.Fragment>
+    <>
       <input
         // eslint-disable-next-line jsx-a11y/no-autofocus
         autoFocus
@@ -121,7 +121,7 @@ function NamedSetNodeEditing(props) {
       >
         Save
       </button>
-    </React.Fragment>
+    </>
   );
 }