+++ src/components/spatial/Spatial.js
@@ -154,30 +154,20 @@ export default class Spatial extends AbstractSelectableComponent {
   }
 
   createRasterLayer() {
-    const source = this.raster;
-    const sourceChannels = {};
-    Object.keys(source).forEach((channel) => {
-      sourceChannels[channel] = source[channel].tileSource;
-    });
-    const { colorValues, sliderValues, channelsOn } = this.props;
-    if (colorValues && sliderValues && channelsOn && sourceChannels) {
-      const channelsLength = Object.keys(sourceChannels).length;
-      if (
-        channelsLength === Object.keys(colorValues).length
-        && channelsLength === Object.keys(sliderValues).length
-        && channelsLength === Object.keys(channelsOn).length
-      ) {
-        const props = {
-          useTiff: sourceChannels[Object.keys(sourceChannels)[0]].includes('tif'),
-          useZarr: sourceChannels[Object.keys(sourceChannels)[0]].includes('zarr'),
-          minZoom: source[Object.keys(sourceChannels)[0]].minZoom,
-          sourceChannels,
-          colorValues,
-          sliderValues,
-          channelsOn,
-        };
-        return new VivViewerLayer(props);
-      }
+    const {
+      colorValues, sliderValues, channelVisibilities, raster,
+    } = this.props;
+    if (colorValues && sliderValues && channelVisibilities && raster) {
+      const { loader } = raster;
+      return new VivViewerLayer({
+        loader,
+        colorValues,
+        sliderValues,
+        channelIsOn: channelVisibilities,
+        onTileError: (err) => {
+          throw err;
+        },
+      });
     }
     return null;
   }