+++ src/components/spatial/SpatialSubscriber.js
@@ -17,7 +17,7 @@ import {
   RASTER_ADD,
   SLIDERS_CHANGE,
   COLORS_CHANGE,
-  CHANNEL_TOGGLE,
+  CHANNEL_VISIBILITY_CHANGE,
 } from '../../events';
 import Spatial from './Spatial';
 
@@ -28,12 +28,16 @@ export default class SpatialSubscriber extends React.Component {
       cells: {},
       selectedCellIds: new Set(),
       cellColors: null,
+      colorValues: null,
+      sliderValues: null,
+      channelVisibilities: null,
       raster: null,
     };
     this.componentWillUnmount = this.componentWillUnmount.bind(this);
   }
 
-  componentWillMount() {
+  // eslint-disable-next-line camelcase
+  UNSAFE_componentWillMount() {
     this.moleculesAddToken = PubSub.subscribe(
       MOLECULES_ADD, this.moleculesAddSubscriber.bind(this),
     );
@@ -55,9 +59,6 @@ export default class SpatialSubscriber extends React.Component {
     this.rasterAddToken = PubSub.subscribe(
       RASTER_ADD, this.rasterAddSubscriber.bind(this),
     );
-    this.slidersChangeToken = PubSub.subscribe(SLIDERS_CHANGE, this.onSlidersChange.bind(this));
-    this.colorsChangeToken = PubSub.subscribe(COLORS_CHANGE, this.onColorsChange.bind(this));
-    this.channelToggleToken = PubSub.subscribe(CHANNEL_TOGGLE, this.onChannelToggle.bind(this));
   }
 
   componentDidMount() {
@@ -75,7 +76,7 @@ export default class SpatialSubscriber extends React.Component {
     PubSub.unsubscribe(this.rasterAddToken);
     PubSub.unsubscribe(this.slidersChangeToken);
     PubSub.unsubscribe(this.colorsChangeToken);
-    PubSub.unsubscribe(this.channelToggleToken);
+    PubSub.unsubscribe(this.channelVisibilityChangeToken);
   }
 
   cellsSelectionSubscriber(msg, cellIds) {
@@ -84,6 +85,19 @@ export default class SpatialSubscriber extends React.Component {
 
   rasterAddSubscriber(msg, raster) {
     this.setState({ raster });
+    const { id } = raster;
+    this.slidersChangeToken = PubSub.subscribe(
+      SLIDERS_CHANGE(id),
+      this.onSlidersChange.bind(this),
+    );
+    this.colorsChangeToken = PubSub.subscribe(
+      COLORS_CHANGE(id),
+      this.onColorsChange.bind(this),
+    );
+    this.channelVisibilityChangeToken = PubSub.subscribe(
+      CHANNEL_VISIBILITY_CHANGE(id),
+      this.onChannelVisibilityChange.bind(this),
+    );
   }
 
   moleculesAddSubscriber(msg, molecules) {
@@ -102,25 +116,20 @@ export default class SpatialSubscriber extends React.Component {
     this.setState({ cellColors });
   }
 
-  onSlidersChange(msg, sliderData) {
-    const sliderValue = { [sliderData.channel]: sliderData.sliderValue };
-    this.setState(prevState => ({ sliderValues: { ...prevState.sliderValues, ...sliderValue } }));
+  onSlidersChange(msg, sliderValues) {
+    this.setState({ sliderValues });
   }
 
-  onColorsChange(msg, rgbData) {
-    const colorValue = { [rgbData.channel]: rgbData.rgb };
-    this.setState(prevState => ({ colorValues: { ...prevState.colorValues, ...colorValue } }));
+  onColorsChange(msg, colorValues) {
+    this.setState({ colorValues });
   }
 
-  onChannelToggle(msg, channelOn) {
-    const channelOnValue = { [channelOn.channel]: channelOn.channelToggle };
-    this.setState(prevState => ({ channelsOn: { ...prevState.channelsOn, ...channelOnValue } }));
+  onChannelVisibilityChange(msg, channelVisibilities) {
+    this.setState({ channelVisibilities });
   }
 
   render() {
-    const {
-      cells, molecules, sliderValues, colorValues, channelsOn,
-    } = this.state;
+    const { cells, molecules } = this.state;
     const { uuid = null, children, removeGridComponent } = this.props;
     const cellsCount = cells ? Object.keys(cells).length : 0;
     const moleculesCount = molecules ? Object.keys(molecules).length : 0;
@@ -156,9 +165,6 @@ export default class SpatialSubscriber extends React.Component {
           clearPleaseWait={
             layerName => PubSub.publish(CLEAR_PLEASE_WAIT, layerName)
           }
-          sliderValues={sliderValues}
-          colorValues={colorValues}
-          channelsOn={channelsOn}
         />
       </TitleInfo>
       /* eslint-enable */