+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 # Changelog
 
-## 0.0.25 - In progress
-### Added
+## 0.1.0 - In progress
+
 - HiGlass integration
+
+## [0.0.25](https://www.npmjs.com/package/vitessce/v/0.0.25) - 2020-03-26
+### Added
+- Removed nwb, added custom webpack configuration.
+- Added glossary.
+- Added theme props, with support for "light" and "dark". Added a URL theme parameter for the demo.
+
+### Changed
+- Upgrade vitessce-image-viewer to 0.1.3 & use data loaders.
+- Rename "LayerPublisher" to "SourcePublisher".
+- Converted the polygon tool to a lasso tool.
+- Loosened assumptions about cell info object properties when rendering cell hover messages to fix a `TypeError` bug.
 
 ## [0.0.24](https://www.npmjs.com/package/vitessce/v/0.0.24) - 2020-03-02
 ### Added